Fix the ultraquickstart instructions
Summary:
The previous instructions (curl ... | bash) wouldn't work, since bash's
stdin would be from curl, and as such it can't read from the keyboard.
Also the failure of `read` wouldn't cause the script to exit.

Tweak the instructions to avoid clobbering stdin, and fail sooner if
`read` can't work for whatever reason.

Test Plan:
Tried:

    bash <(cat install)

And it ran and prompted me with questions.

Also had @kevinliu give it a try.

Reviewers: #developers, kevinliu

Reviewed By: kevinliu

Subscribers: kevinliu

Differential Revision: https://phabricator.paperg.com/D219