Print nice error on failed old key deactivation
Summary:
Rarely, the old key won't be deactivated. This could be due to AWS being
especially slow to converge.

When that happens, print an error explaining what needs to be done to
fix it. Also increase the time awsudo will keep retrying -- maybe that
will make this failure even more rare. Resolves T561.

Test Plan:
I was never able to reproduce the error reported in T561, so instead I
altered the code so that it would always fail by trying to deactivate a
key that does not exist. Read the error.

Put the code back to normal and verified that it still works normally.

Reviewers: #developers, kian

Reviewed By: #developers, kian

Subscribers: kian

Maniphest Tasks: T561

Differential Revision: https://phabricator.paperg.com/D365